+++ iodine-bukkit/src/main/java/hu/trigary/iodine/bukkit/gui/element/CheckboxGuiElementImpl.java
@@ -28,7 +28,7 @@ public class CheckboxGuiElementImpl extends GuiElementImpl<CheckboxGuiElement> i
 	 * @param internalId the internal ID of this element
 	 * @param id the API-friendly ID of this element
 	 */
-	public CheckboxGuiElementImpl(@NotNull GuiBaseImpl<?> gui, short internalId, @NotNull Object id) {
+	public CheckboxGuiElementImpl(@NotNull GuiBaseImpl<?> gui, int internalId, @NotNull Object id) {
 		super(gui, GuiElementType.CHECKBOX, internalId, id);
 	}
