+++ iodine-bukkit/src/main/java/hu/trigary/iodine/bukkit/gui/element/RadioButtonGuiElementImpl.java
@@ -32,7 +32,7 @@ public class RadioButtonGuiElementImpl extends GuiElementImpl<RadioButtonGuiElem
 	 * @param internalId the internal ID of this element
 	 * @param id the API-friendly ID of this element
 	 */
-	public RadioButtonGuiElementImpl(@NotNull GuiBaseImpl<?> gui, short internalId, @NotNull Object id) {
+	public RadioButtonGuiElementImpl(@NotNull GuiBaseImpl<?> gui, int internalId, @NotNull Object id) {
 		super(gui, GuiElementType.RADIO_BUTTON, internalId, id);
 		setGroupId(0);
 	}
